refactor: add vite-plugin-html. Delete updateHtml related logic
