+++ CHANGELOG.zh_CN.md
@@ -1,4 +1,17 @@
-## # 2.0.0-rc.5 (2020-10-26)
+## Wip
+
+### 🎫 Chores
+
+- 升级 vite 版本为`v1.0.0.rc8`
+- vite.config.ts 内部 plugins 抽取
+- build 目录结构调整
+- 依赖更新
+
+### ✨ Refactor
+
+- 独立出`vite-plugin-html`,并修改相关插入 html 的逻辑
+
+## 2.0.0-rc.5 (2020-10-26)
 
 ### ✨ Features
