+++ index.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
+    <%= viteHtmlPluginOptions.hmScript %>
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
     <meta name="renderer" content="webkit" />
@@ -10,6 +11,7 @@
     />
     <title></title>
     <link rel="icon" href="/favicon.ico" />
+    <%= viteHtmlPluginOptions.injectConfig %>
   </head>
   <body>
     <div id="app">
