+++ package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "bootstrap": "yarn install",
     "serve": "esno ./build/script/preserve.ts && cross-env NODE_ENV=development vite",
-    "build": "cross-env NODE_ENV=production vite build && esno ./build/script/postBuild.ts",
+    "build": "rimraf dist && cross-env NODE_ENV=production vite build && esno ./build/script/postBuild.ts",
     "build:site": "cross-env SITE=true npm run build ",
     "build:no-cache": "yarn  clean:cache && npm run build",
     "report": "cross-env REPORT=true npm run build ",
@@ -48,7 +48,6 @@
     "@purge-icons/generated": "^0.4.1",
     "@types/echarts": "^4.8.3",
     "@types/fs-extra": "^9.0.2",
-    "@types/html-minifier": "^4.0.0",
     "@types/koa-static": "^4.0.1",
     "@types/lodash-es": "^4.17.3",
     "@types/mockjs": "^1.0.3",
@@ -72,7 +71,6 @@
     "eslint-plugin-vue": "^7.1.0",
     "esno": "^0.2.4",
     "fs-extra": "^9.0.1",
-    "html-minifier": "^4.0.0",
     "husky": "^4.3.0",
     "koa-static": "^5.0.0",
     "less": "^3.12.2",
@@ -90,6 +88,7 @@
     "tasksfile": "^5.1.1",
     "ts-node": "^9.0.0",
     "typescript": "^4.0.5",
+    "vite-plugin-html": "^1.0.0-beta.2",
     "vite-plugin-mock": "^1.0.4",
     "vite-plugin-purge-icons": "^0.4.4",
     "vue-eslint-parser": "^7.1.1",
